zodbdump: Switch to using qq from pygolang
I originally added escapeqq as part of 75c03368 (zodbdump: Start to
stabilize output format) with the task for this utility to quote string
into valid "..." string always quoted with ".

This utility was later copied to pygolang:

	https://lab.nexedi.com/kirr/pygolang/commit/afa46cf5#aafe23bd3b5a060d8afd59367774e219700c7064_0_22

and then further improved there to work under both Python2 and Python3
and to not escape printable UTF-8 characters:

	https://lab.nexedi.com/kirr/pygolang/commit/02dddb97

So stop the duplication and simply switch to the better version.