zodbdump: support --pretty option with a format to show pickles disassembly
Showing pickle disassembly can sometimes be useful to analyse details of
the pickle content. We realized that in some data structures used in
ERP5 the same string was saved multiple times in the same pickle and by
using the exact same string (ie. for which `s1 is s2` is True), the
pickle will have the string only once and pickles are a bit smaller. For
more reference, the context was
https://lab.nexedi.com/nexedi/erp5/merge_requests/1560#note_154825

This introduces a new --pretty option that we will be able to extend
later with more output formats.

Co-authored-by: Kirill Smelkov <kirr@nexedi.com>