test/gen_testdata: Prevent object with OID=1 to be deleted
Previously it was like this by luck, but recent changes shuffled
generated output and one of delete selection started to point to OID 1:

    (zdev+py39.env) kirr@deca:~/src/wendelin/z/zodbtools/zodbtools/test/testdata$ grep '^obj .* delete' 1.zdump.zpickledis.ok
    obj 0000000000000001 delete
    obj 0000000000000002 delete

However NEO/go tests depend on object with OID=1 to be present:

https://lab.nexedi.com/kirr/neo/-/blob/1ad088c8/go/neo/neo_test.go#L216-219

and it breaks there if here we let that object with OID=1 to be deleted.

-> Prevent ZODB/go breakage by explicitly making sure that OID=1 stays
present under any conditions.

The actual places where gen_testdata.py from zodbtools is used in
ZODB/go and Wendelin.core are e.g. here:

https://lab.nexedi.com/kirr/neo/-/blob/f7776fc1/go/zodb/storage/fs1/py/gen-testdata
https://lab.nexedi.com/nexedi/wendelin.core/-/blob/07087ec8/wcfs/internal/zdata/testdata/zblk_test_gen.py