zodbanalyze: tests: Verify CSV mode via golden output generated during gen_testdata
Instead of putting expected output into test code.

The reason for this change is that soon, with many kinds of generated
ZODB databases, analyze output will be different for each kind, and the
only practical way to make sure that analyze works correctly is to have
good expected output for each kind.

We already do this way for zodbdump. Start doing the same for
zodbanalyze as a preparatory step as well.