commit: test: Verify it with both FileStorage and ZEO
There is currently one problem with `zodb commit` that exhibits it only
with ZEO. And it could be triggered only via commandline.

As a preparatory step for the fix add infrastructure to test zodbcommit
with both FileStorage and ZEO via zsrv fixture.

There is currently no breakage - test_commit continues to work ok. In
the next patch we will add test for commandline handling of `zodb
commit` and fix the issue.

P.S. Support code for testing wrt ZEO is based on wendelin.core :

https://lab.nexedi.com/nexedi/wendelin.core/-/blob/07087ec8/lib/testing.py#L230-308

Maybe in the future it would be a good idea to switch wendelin.core to
use zodbtools-provided ZSrv fixture.