tests: Fix ztestdata-related tests on ZODB4
In bf772ce0 I introduced ztestdata fixture and we added several tests
that use this fixture instead of only zext fixture. But contrary to zext
I missed to corresponding xfail if underlying ZODB is ZODB4 and does not
support access to raw extension bytes. As the result several tests
started to fail with ZODB4, because on such ZODB extension is only
dumped via heuristic and on best-effort basis and that cannot match
original extension data bit to bit exactly:

https://lab.nexedi.com/nexedi/zodbtools/-/blob/36cba871/zodbtools/zodbdump.py#L93-102

One example failure is here:

https://erp5js.nexedi.net/#/test_result_module/20241017-58899E0D/2

-> Fix that by xfailing in both zext and ztestdata fixtures
consistently.

Co-authored-by: Jérome Perrin <jerome@nexedi.com>
/reviewed-on https://lab.nexedi.com/nexedi/zodbtools/-/merge_requests/31 and https://lab.nexedi.com/nexedi/zodbtools/-/merge_requests/32